Complete car request flow (create a message in queue then consume that message and broadcast to all suitable drivers)
